refactor(notify): split notify + prompt to satisfy module-health cap
Follow-up to #3419. That change grew src/teatree/core/notify.py (501→528 LOC)
and src/teatree/agents/prompt.py (~490→534 LOC) past the 500-LOC module-health
cap. module-health-gate is not a branch-protection required check so #3419
merged, but main now carries two over-cap files. Split each by concern (no
behavior change):

- notify.py: extract the DM-target resolvers (resolve_user_id,
  resolve_user_channel, _sole_overlay_field) into a new
  teatree.core.notify_targets module — the "who/which channel does the bot DM"
  concern. Re-exported from teatree.core.notify so every
  `from teatree.core.notify import resolve_user_id` call site is unchanged.

- prompt.py: extract the coding-builder dispatch contract (the forced-skill-load
  + behavior-preservation + verify directive, _coding_phase_directive and its
  _stack_* helpers, plus _VERIFY_GATES_COMMAND / _HEARTBEAT_DM_LINES) into a new
  teatree.agents.coding_prompt module. One-directional: prompt imports from
  coding_prompt, never the reverse.

Both source files are back under the 500-LOC cap; the two new modules are small
(≤2 public functions each). Pure code-move refactor — existing tests pass
unchanged via the re-exports.
