+++ src/teatree/core/notify.py
@@ -31,11 +31,12 @@ import re
 
 from django.db import DatabaseError, IntegrityError, transaction
 
-from teatree.config import get_effective_settings, load_config
+from teatree.config import get_effective_settings
 from teatree.core.backend_factory import messaging_from_overlay
 from teatree.core.backend_protocols import MessagingBackend
 from teatree.core.modelkit.notify_policy import NotifyAudience
 from teatree.core.models import BotPing, DeliveryClaim, OutboundClaim
+from teatree.core.notify_targets import resolve_user_channel, resolve_user_id
 from teatree.core.send_proxy import SendChannel, SendRequest, route_send
 from teatree.core.session_identity import current_session_id
 from teatree.slack_mrkdwn import normalize_slack_message, slack_linkify
@@ -265,8 +266,8 @@ def _maybe_stamp_answered(*, idempotency_key: str, answering_slack_ts: str) -> N
         if match is None:
             return
         ts = match.group(1)
-    # Deferred import (mirrors ``resolve_user_id`` / ``maybe_linkify``
-    # in this module): the answer-stamp is an opt-in side path; keeping
+    # Deferred import (mirrors ``maybe_linkify`` in this module): the
+    # answer-stamp is an opt-in side path; keeping
     # the model import out of ``teatree.core.notify`` import time avoids
     # perturbing the module-import graph that the on-behalf gate and
     # notify suites rely on.
@@ -430,82 +431,6 @@ def _feature_enabled() -> bool:
     return bool(getattr(settings_, "notify_user_via_bot", True))
 
 
-def resolve_user_id() -> str:
-    """Resolve the Slack user id to DM (overlay override → global → sole overlay → empty).
-
-    The per-overlay id comes from the DB overlays registry (still injected into
-    ``load_config().raw["overlays"]``); the GLOBAL fallback reads the DB-home
-    ``slack_user_id`` setting so every routing path agrees on the same order.
-
-    The final ``sole overlay`` tier is the env-independent fallback that mirrors
-    :func:`teatree.core.backend_factory.messaging_from_overlay` — the headless
-    worker that DMs the owner does NOT export ``T3_OVERLAY_NAME``, so the
-    overlay-scoped tier is skipped, and a fresh box carries no GLOBAL setting.
-    When exactly one overlay is registered there is no ambiguity, so its own
-    ``slack_user_id`` resolves without depending on the env var being set.
-    """
-    from teatree.config import cold_reader  # noqa: PLC0415 — deferred: call-time import, kept lazy
-
-    cfg = load_config().raw
-    overlay_name = os.environ.get("T3_OVERLAY_NAME", "")
-    overlays = cfg.get("overlays") or {}
-    if overlay_name and isinstance(overlays.get(overlay_name), dict):
-        user_id = overlays[overlay_name].get("slack_user_id", "")
-        if user_id:
-            return str(user_id)
-    global_id = cold_reader.str_setting("slack_user_id", default="")
-    if global_id:
-        return global_id
-    return _sole_overlay_field(overlays, "slack_user_id")
-
-
-def resolve_user_channel() -> str:
-    """Resolve the Slack DM channel id the user reads (overlay → global → sole overlay → empty).
-
-    The canonical resolver for the ``slack_user_channel`` config key,
-    walking the SAME overlay→global→sole-overlay→empty order :func:`resolve_user_id`
-    uses for ``slack_user_id``. Both DM-channel call sites (the bot→user
-    DM path and the live-post-approval CLI verifier) consult this single
-    helper, so a change to the resolution order can never drift between
-    two private copies (the config-trap the #126 redesign closes).
-
-    An empty return means no channel is configured; the caller treats it
-    as "open a DM to the resolved user_id" rather than pinning to a
-    specific ``D...`` channel.
-    """
-    from teatree.config import cold_reader  # noqa: PLC0415 — deferred: call-time import, kept lazy
-
-    cfg = load_config().raw
-    overlay_name = os.environ.get("T3_OVERLAY_NAME", "")
-    overlays = cfg.get("overlays") or {}
-    if overlay_name and isinstance(overlays.get(overlay_name), dict):
-        channel = overlays[overlay_name].get("slack_user_channel", "")
-        if channel:
-            return str(channel)
-    global_channel = cold_reader.str_setting("slack_user_channel", default="")
-    if global_channel:
-        return global_channel
-    return _sole_overlay_field(overlays, "slack_user_channel")
-
-
-def _sole_overlay_field(overlays: dict, key: str) -> str:
-    """Return the sole registered overlay's ``key`` value, or ``""``.
-
-    Env-independent fallback for the DM-target resolvers: when the active
-    overlay is ambiguous (``T3_OVERLAY_NAME`` unset in the headless worker)
-    and no GLOBAL setting is configured, a single registered overlay is
-    unambiguous — its own value is the right target. Returns ``""`` when
-    zero or more than one overlay is registered (ambiguous), so a
-    multi-overlay box never silently picks the wrong owner.
-    """
-    if len(overlays) != 1:
-        return ""
-    entry = next(iter(overlays.values()))
-    if not isinstance(entry, dict):
-        return ""
-    return str(entry.get(key, "") or "")
-
-
 def maybe_linkify(text: str) -> str:
     """Apply :func:`slack_linkify` using the active overlay's resolvers, if any.
 
@@ -653,4 +578,10 @@ def drain_undelivered_notifies(
     return delivered, len(rows)
 
 
-__all__ = ["NotifyKind", "drain_undelivered_notifies", "notify_user"]
+__all__ = [
+    "NotifyKind",
+    "drain_undelivered_notifies",
+    "notify_user",
+    "resolve_user_channel",
+    "resolve_user_id",
+]
